+++ skt/__init__.py
@@ -12,11 +12,6 @@
 # along with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 """Functions used by multiple parts of skt."""
-from email.errors import HeaderParseError
-import email.header
-import email.parser
-import re
-
 import requests
 
 from skt.misc import join_with_slash
@@ -47,35 +42,3 @@ def get_patch_mbox(url):
                         (url, response.status_code))
 
     return response.content
-
-
-def get_patch_name(content):
-    """
-    Retrieve patch name from 'Subject' header from the mbox string
-    representing a patch.
-
-    Args:
-        content: String representing patch mbox
-
-    Returns:
-        Name of the patch. <SUBJECT MISSING> is returned if no subject is
-        found, and <SUBJECT ENCODING INVALID> if header decoding fails.
-    """
-    headers = email.parser.Parser().parsestr(content, True)
-    subject = headers['Subject']
-    if not subject:
-        # Emails return None if the header is not found so use a stub subject
-        # instead of it
-        return '<SUBJECT MISSING>'
-
-    # Remove header folding
-    subject = re.sub(r'\r?\n[ \t]', ' ', subject)
-
-    try:
-        # decode_header() returns a list of tuples (value, charset)
-        decoded = [value for value, _ in email.header.decode_header(subject)]
-    except HeaderParseError:
-        # We can't parse the original subject so use a stub one instead
-        return '<SUBJECT ENCODING INVALID>'
-
-    return ''.join(decoded)
