+++ skt/reporter.py
@@ -25,7 +25,7 @@ import StringIO
 import requests
 
 import skt
-from skt.misc import join_with_slash
+from skt.misc import join_with_slash, get_patch_name
 import skt.runner
 
 
@@ -333,7 +333,7 @@ class Reporter(object):
         if self.cfg.get("patchworks"):
             for purl in self.cfg.get("patchworks"):
                 patch_mbox = skt.get_patch_mbox(purl)
-                patchname = skt.get_patch_name(patch_mbox)
+                patchname = get_patch_name(patch_mbox)
                 mergedata['patchwork'].append((purl, patchname))
 
         return mergedata
