+++ tests/test_misc.py
@@ -12,7 +12,10 @@
 # this program; if not, write to the Free Software Foundation, Inc., 51
 # Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 """Test cases for misc.py."""
+from email.errors import HeaderParseError
 import unittest
+
+import mock
 
 import skt.misc
 
@@ -47,3 +50,43 @@ class TestIndependent(unittest.TestCase):
         suffix = "part"
         self.assertEqual("http://url.com/part",
                          skt.misc.join_with_slash(base, suffix))
+
+    def test_nonexistent_patch_subject(self):
+        """Ensure get_patch_name() handles nonexistent 'Subject' in mbox."""
+        mbox_body = 'nothing useful here'
+        self.assertEqual('<SUBJECT MISSING>',
+                         skt.misc.get_patch_name(mbox_body))
+
+    def test_ok_patch_subject(self):
+        """Ensure get_patch_name() returns correct 'Subject' if present."""
+        mbox_body = 'From Test Thu May 2 17:49:51 2018\nSubject: GOOD SUBJECT'
+        self.assertEqual('GOOD SUBJECT',
+                         skt.misc.get_patch_name(mbox_body))
+
+    def test_encoded_patch_subject(self):
+        """Ensure get_patch_name() correctly decodes UTF-8 'Subject'."""
+        mbox_body = ('From Test Thu May 2 17:49:51 2018\n'
+                     'Subject: =?utf-8?q?=5BTEST=5D?=')
+        self.assertEqual('[TEST]', skt.misc.get_patch_name(mbox_body))
+
+    @mock.patch('email.header.decode_header')
+    @mock.patch('email.parser.Parser.parsestr')
+    def test_header_parse_failure(self, mock_parsestr, mock_decode_header):
+        """Ensure get_patch_name() handles a parsing failure."""
+        mock_parsestr.return_value = {'Subject': "Testing"}
+        mock_decode_header.side_effect = HeaderParseError('Fail')
+        result = skt.misc.get_patch_name('')
+
+        self.assertEqual('<SUBJECT ENCODING INVALID>', result)
+
+    def test_multipart_encoded_subject(self):
+        """
+        Ensure get_patch_name() correctly decodes multipart encoding
+        of 'Subject'.
+        """
+        mbox_body = ('From Test Thu May 2 17:49:51 2018\nSubject: '
+                     '=?ISO-8859-1?B?SWYgeW91IGNhbiByZWFkIHRoaXMgeW8=?=\n'
+                     '    =?ISO-8859-2?B?dSB1bmRlcnN0YW5kIHRoZSBleGFtcGxlLg'
+                     '==?=')
+        self.assertEqual('If you can read this you understand the example.',
+                         skt.misc.get_patch_name(mbox_body))
